Split AppConnector into NiceBaseConnector
